docs: continue rename
