+++ src/repo_review/families.py
@@ -18,9 +18,7 @@ class Family(typing.TypedDict, total=False):
 def collect_families() -> dict[str, Family]:
     return {
         name: family
-        for ep in importlib.metadata.entry_points(
-            group="repo_review.families"
-        )
+        for ep in importlib.metadata.entry_points(group="repo_review.families")
         for name, family in ep.load()().items()
     }
